SamsonovTourList/requirements.txt
Django>=3.0.3
django-import-export>=*
djangorestframework>=*
graphene_django>=*
mysql>=*
django-cors-headers>=*